@@ -21,31 +21,25 @@ public class AnimationsActivity extends Activity {
 
         final View fab = findViewById(R.id.fab);
         Button button = (Button) findViewById(R.id.button);
-        button.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                if (fabVisibility != View.VISIBLE) {
-                    fab.setVisibility(View.VISIBLE);
-                    fabVisibility = View.VISIBLE;
-                } else {
-                    fab.setVisibility(View.INVISIBLE);
-                    fabVisibility = View.INVISIBLE;
-                }
+        button.setOnClickListener(view -> {
+            if (fabVisibility != View.VISIBLE) {
+                fab.setVisibility(View.VISIBLE);
+                fabVisibility = View.VISIBLE;
+            } else {
+                fab.setVisibility(View.INVISIBLE);
+                fabVisibility = View.INVISIBLE;
             }
         });
 
         final View button2 = findViewById(R.id.button2);
         Button button3 = (Button) findViewById(R.id.button3);
-        button3.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                if (buttonVisibility != View.VISIBLE) {
-                    button2.setVisibility(View.VISIBLE);
-                    buttonVisibility = View.VISIBLE;
-                } else {
-                    button2.setVisibility(View.INVISIBLE);
-                    buttonVisibility = View.INVISIBLE;
-                }
+        button3.setOnClickListener(view -> {
+            if (buttonVisibility != View.VISIBLE) {
+                button2.setVisibility(View.VISIBLE);
+                buttonVisibility = View.VISIBLE;
+            } else {
+                button2.setVisibility(View.INVISIBLE);
+                buttonVisibility = View.INVISIBLE;
             }
         });
     }
