@@ -30,36 +30,16 @@ public class CircularProgressActivity extends Activity {
                 }
             }
         };
-        findViewById(R.id.button1).setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                progress1.setVisibility(progress1.getVisibility() == View.VISIBLE ? View.INVISIBLE : View.VISIBLE);
-            }
-        });
+        findViewById(R.id.button1).setOnClickListener(view -> progress1.setVisibility(progress1.getVisibility() == View.VISIBLE ? View.INVISIBLE : View.VISIBLE));
 
         final View progress2 = findViewById(R.id.progress2);
-        findViewById(R.id.button2).setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                progress2.setVisibility(progress2.getVisibility() == View.VISIBLE ? View.INVISIBLE : View.VISIBLE);
-            }
-        });
+        findViewById(R.id.button2).setOnClickListener(view -> progress2.setVisibility(progress2.getVisibility() == View.VISIBLE ? View.INVISIBLE : View.VISIBLE));
 
         final View progress3 = findViewById(R.id.progress3);
-        findViewById(R.id.button3).setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                progress3.setVisibility(progress3.getVisibility() == View.VISIBLE ? View.INVISIBLE : View.VISIBLE);
-            }
-        });
+        findViewById(R.id.button3).setOnClickListener(view -> progress3.setVisibility(progress3.getVisibility() == View.VISIBLE ? View.INVISIBLE : View.VISIBLE));
 
         final View progress4 = findViewById(R.id.progress4);
-        findViewById(R.id.button4).setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                progress4.setVisibility(progress4.getVisibility() == View.VISIBLE ? View.INVISIBLE : View.VISIBLE);
-            }
-        });
+        findViewById(R.id.button4).setOnClickListener(view -> progress4.setVisibility(progress4.getVisibility() == View.VISIBLE ? View.INVISIBLE : View.VISIBLE));
 
         handler.postDelayed(runnable, 10);
     }
