@@ -2,7 +2,6 @@ package tk.zielony.carbonsamples.widget;
 
 import android.app.Activity;
 import android.os.Bundle;
-import android.view.View;
 import android.view.ViewGroup;
 
 import carbon.widget.CheckBox;
@@ -27,28 +26,17 @@ public class SnackbarActivity extends Activity {
         final CheckBox pushCheckBox = (CheckBox) findViewById(R.id.push);
         final FloatingActionButton fab = (FloatingActionButton) findViewById(R.id.fab);
 
-        findViewById(R.id.button).setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                final Snackbar snackbar = new Snackbar(SnackbarActivity.this, "Hello world!", "dismiss", infiniteCheckBox.isChecked() ? Snackbar.INFINITE : getResources().getInteger(R.integer.carbon_snackbarDuration));
-                snackbar.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        snackbar.dismiss();
-                    }
-                });
-                snackbar.setStyle(floatingCheckBox.isChecked() ? Snackbar.Style.Floating : Snackbar.Style.Docked);
-                snackbar.setTapOutsideToDismissEnabled(tapCheckBox.isChecked());
-                snackbar.setSwipeToDismissEnabled(swipeCheckBox.isChecked());
-                if (pushCheckBox.isChecked())
-                    snackbar.addPushedView(fab);
-                snackbar.show((ViewGroup) getWindow().getDecorView().getRootView());
-                snackbar.setOnDismissListener(new Snackbar.OnDismissListener() {
-                    @Override
-                    public void onDismiss() {
-                    }
-                });
-            }
+        findViewById(R.id.button).setOnClickListener(view -> {
+            final Snackbar snackbar = new Snackbar(SnackbarActivity.this, "Hello world!", "dismiss", infiniteCheckBox.isChecked() ? Snackbar.INFINITE : getResources().getInteger(R.integer.carbon_snackbarDuration));
+            snackbar.setOnClickListener(v -> snackbar.dismiss());
+            snackbar.setStyle(floatingCheckBox.isChecked() ? Snackbar.Style.Floating : Snackbar.Style.Docked);
+            snackbar.setTapOutsideToDismissEnabled(tapCheckBox.isChecked());
+            snackbar.setSwipeToDismissEnabled(swipeCheckBox.isChecked());
+            if (pushCheckBox.isChecked())
+                snackbar.addPushedView(fab);
+            snackbar.show((ViewGroup) getWindow().getDecorView().getRootView());
+            snackbar.setOnDismissListener(() -> {
+            });
         });
 
         Snackbar.clearQueue();
